[#2] Fix for previous redis client compatibility issue
Due to the change in the redis client `with_scores -> withscores` we
applied a fix to make it compatible regardless of which arg spelling was
used. That change inadvertently tossed the other valid options that
should have been passed to the sorted set methods. This change undoes
that capatibility fix and simply passes the options through to the redis
client (as it should and did do previously).