[NOJIRA] Raise an exception for boolean and nil values
w/ @sebastien-coavoux

HashKey objects are stored as hashes in redis. When setting a hash key,
we choose to raise an exception if the value is a boolean or nil.

e.g. `hashkey["test"] = true`

This is a conservative approach to avoid unexpected behavior when
reading the value back. Unlike object serialization with JSON or YAML,
redis hash key values are always strings. Without a more nuanced way to
configure types, there's no way for Famila to tell the difference
between "true" as in true and "true" as in the literal string "true". We
may revisit in future versions.