[NOJIRA] Version bump: 0.10.1
Release as a patch update even though we've added a change in behvaiour (raising an exceptioon when setting potentally ambiguous value types in a redis hash). Mainly b/c prior to this change, the benviour was indeterminate. A boolean true value would appear to be stored as a boolean but later read back as a string (since we don't do explicit conversion).

Signed-off-by: delano <delano@cpan.org>