[#20240824] Support for skipping expiration update on save
Optimize save operation with optional expiration update

This change enhances the flexibility of the save process by:
- Introducing an `update_expiration` parameter in `save` method
- Propagating this option to `commit_fields`
- Conditionally calling `update_expiration` based on the parameter
- Updating debug logs to reflect the new behavior

This change allows for updating the state of an object without affecting
its original expiration time. e.g. when a metadata key is marked as
viewed, we still want it to expire at the same time as before.