[#V1] Add has_relations? method manually
In lieu of a centralized approach to managing relations methods between
Horreum (which can have them) and RedisType (which can't) classes, we
take the KISS approach and simply add a method with the same name to
both.

NOTE: has_relations? only ever returns true for Horreum classes and only
if the subclass has at least one relation defined. We use this in
update_expiration for example to update the ttls on related objects
(only if they specified a ttl when they were defined).