Remove remove_from_all_participations
Remove for being poorly implemented. It was repetitive and a slog
to debug. It'll come back in a cleaner for after the rest of the
module is in ship shape. It's cool how it would automatically remove
instances from `sorted_set` etc data type fields, not just indexed
and participations.