Simplify automatic instance tracking
Implement automatic instance tracking using built-in class_sorted_set
to avoid having to enable relationship feature for all horreum model
classes. We can add the instance to the sorted set with every
successful save of the main object, using the current time as the
score and leave it to Valkey/Redus to sort out whether it is truly
adding or updating the score for an existing instance.