BREAKING: SortedSet.add argument order swapped
The redis commands for zsets have the score first, followed by the
value. Here that prevents us from being able to default the score to
the current time (it's the most common way to score, although not the
only way). This has the positive impact that all collection DataType
classes can now have the same API for the add method: .add(value)