Use horreum management methods
Relationship feature functionality is only available to horreum
model classes so it makes sense to re-use existing core functionality
rather than reinventing the self-sealing stem bolts.

Also fixes a testcase issue where the test objects weren't being
persisted to the database. These two issues unfortunately just
happened to cancel each other out in terms of visibility and not
having caught either one of them earlier. It indicates a lack of
test coverage as well which we'll be adding in with this PR as well.