Add new integration tryouts (first pass)
Each one includes a verbose amount of inline documentation describing
the intention behind the test coverage. We'll use this while working
on the test cases and then pare it back at the tail end.

Full save/load/find cycles through relationship queries. There are
five other files that we'll include in a separate branch and PR. They
get into bugs and/or behaviour that we want to revisit.

NOTE: There are two testcases in these tryouts that are marked as
intentionally failing (i.e with `#=<>` prefix. Should we return all
fields as strings like we always have or do we want to support basic
JSON types like Integer? Food for thought.