[#157] Rename Management.create to create!
Without the bang, a general create method is a footgun. We want to
know without any doubt that the object was created or not. So the
rename is to intentionally "raise awareness" that the contract when
instantiating and persisting a horreum model in this way is that
you'll either get an instance of the model or a raised error.