Update save and save_if_not_exists! to use shared logic
`save_if_not_exists` doesn't call save because of the gap between
checking existence and persisting the data. We can't check for existence
inside the transaction because commands are queued and not executed
until EXEC is called (if you try you get a Redis::Future object). So
here we use a WATCH + MULTI/EXEC pattern to fail the transaction if the
key is created (or modified in any way) to avoid silent data
corruption♀︎.

♀︎ Additional note about WATCH + MULTI/EXEC in Valkey/Redis or any two
step existence check in any database: although it is more cautious, it
is not atomic. The only way to do that is if the database process can
determine itself whether the record already exists or not. For
Valkey/Redis, that means writing the lua to do that.